Device Layer/requirements.txt
datetime
timedelta
randomtimestamp
kafka-python
influx_line_protocol